vapoursynth: Don't remove existing libs unless to prevent compilation

+++ build/media-suite_compile.sh
@@ -1393,7 +1393,7 @@ if [[ $xpcomp = "n" && $mpv != "n" ]] && pkg-config --exists "libavcodec libavut
                 -e 's;lvapoursynth-script;lvsscript;' \
                 -e '/Libs.private/ d' \
                 > "$LOCALDESTDIR"/lib/pkgconfig/vapoursynth-script.pc
-        else
+        elif [[ x"$vsprefix" = "x" ]]
             rm -f "$LOCALDESTDIR"/lib/pkgconfig/vapoursynth{,-script}.pc
             rm -f "$LOCALDESTDIR"/lib/{vapoursynth,vsscript}.lib
             rm -rf "$LOCALDESTDIR"/include/vapoursynth
